# openapi.yaml
...
openapi: 3.0.3
info:
  title: HyperSDK API
  description: |
    VM migration and export platform providing unified VM export across 9 cloud/virtualization platforms with carbon-aware scheduling.

    Supports: vSphere, Nutanix AHV, AWS, Azure, GCP, Hyper-V, OCI, OpenStack, Alibaba Cloud, Proxmox

    Features:
    - Multi-cloud VM export
    - Carbon-aware scheduling (30-50% carbon reduction)
    - Real-time job monitoring
    - Scheduled backups
  version: 2.0.0
  contact:
    name: HyperSDK
    url: https://github.com/zyvorai/transiva
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.transiva.example.com
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /schedules:
    get:
      summary: List schedules
      description: List all scheduled jobs
      tags: [Schedules]
      responses:
        '200':
          description: List of scheduled jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduledJob'
    post:
      summary: Create schedule
      description: Create a new scheduled job
      tags: [Schedules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledJob'
      responses:
        '200':
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledJob'
  /schedules/{scheduleId}:
    get:
      summary: Get schedule
      description: Get details of a specific schedule
      tags: [Schedules]
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Schedule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledJob'
    put:
      summary: Update schedule
      description: Update an existing schedule
      tags: [Schedules]
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledJob'
      responses:
        '200':
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledJob'
    delete:
      summary: Delete schedule
      description: Delete a scheduled job
      tags: [Schedules]
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Schedule deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token obtained from /api/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
    JobDefinition:
      type: object
      required:
        - vm_path
      description: |
        Job definition for VM export. Supports carbon-aware scheduling via metadata fields.

        For carbon-aware scheduling, set metadata fields:
        - carbon_aware: true
        - carbon_zone: "US-CAL-CISO"
        - carbon_max_intensity: 200.0
        - carbon_max_delay: 14400000000000 (4 hours in nanoseconds)

        Job will be automatically delayed if grid carbon intensity exceeds threshold.
      properties:
        id:
          type: string
          description: Unique job identifier
        name:
          type: string
          description: Human-readable job name
        vm_path:
          type: string
          description: Path to the VM to export
          example: /Datacenter/vm/my-vm
        output_path:
          type: string
          description: Output path for exported files (CLI/YAML)
        output_dir:
          type: string
          description: Output directory for exported files (API)
        vcenter_url:
          type: string
          description: vCenter URL (CLI/YAML)
          example: vcenter.example.com
        vcenter:
          $ref: '#/components/schemas/VCenterConfig'
        username:
          type: string
          description: vCenter username
        datacenter:
          type: string
          description: Datacenter name
        format:
          type: string
          description: Export format
          enum: [qcow2, raw, vmdk, ova, ovf]
          default: ovf
        export_method:
          type: string
          description: Export method to use
          enum: [ctl, govc, ovftool, web, ""]
        method:
          type: string
          description: Alias for export_method (web API compatibility)
        compress:
          type: boolean
          description: Enable GZIP compression
          default: false
        thin:
          type: boolean
          description: Use thin provisioning
          default: false
        insecure:
          type: boolean
          description: Skip TLS certificate verification
          default: false
        options:
          $ref: '#/components/schemas/ExportOptions'
        metadata:
          type: object
          description: |
            Custom metadata for job. Used for carbon-aware scheduling.

            Carbon-aware fields:
            - carbon_aware (bool): Enable carbon-aware scheduling
            - carbon_zone (string): Carbon zone ID (e.g., "US-CAL-CISO")
            - carbon_max_intensity (float): Max carbon intensity threshold in gCO2/kWh
            - carbon_max_delay (int64): Max delay in nanoseconds
          additionalProperties: true
          example:
            carbon_aware: true
            carbon_zone: US-CAL-CISO
            carbon_max_intensity: 200.0
            carbon_max_delay: 14400000000000
        created_at:
          type: string
          format: date-time
    ExportOptions:
      type: object
      properties:
        parallel_downloads:
          type: integer
          description: Number of parallel downloads
          default: 4
          minimum: 1
          maximum: 32
        remove_cdrom:
          type: boolean
          description: Remove CD-ROM devices before export
          default: false
        show_individual_progress:
          type: boolean
          description: Show progress for each file
          default: false
        enable_pipeline:
          type: boolean
          description: Enable h2kvm pipeline integration
        h2kvm_path:
          type: string
          description: Path to h2kvm executable
        hyper2kvm_path:
          type: string
          deprecated: true
          description: 'Deprecated: use h2kvm_path.'
        pipeline_inspect:
          type: boolean
        pipeline_fix:
          type: boolean
        pipeline_convert:
          type: boolean
        pipeline_validate:
          type: boolean
        pipeline_compress:
          type: boolean
        compress_level:
          type: integer
          minimum: 1
          maximum: 9
        libvirt_integration:
          type: boolean
        libvirt_uri:
          type: string
        libvirt_autostart:
          type: boolean
        libvirt_bridge:
          type: string
        libvirt_pool:
          type: string
    VCenterConfig:
      type: object
      required:
        - server
        - username
        - password
      properties:
        server:
          type: string
          description: vCenter server address
          example: vcenter.example.com
        username:
          type: string
          description: vCenter username
          example: administrator@vsphere.local
        password:
          type: string
          format: password
          description: vCenter password
        insecure:
          type: boolean
          description: Skip TLS certificate verification
          default: false
...
    ScheduledJob:
      type: object
      required:
        - name
        - schedule
        - job_template
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        schedule:
          type: string
          description: Cron format schedule
          example: "0 2 * * *"
        job_template:
          $ref: '#/components/schemas/JobDefinition'
        enabled:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        next_run:
          type: string
          format: date-time
        last_run:
          type: string
          format: date-time
        run_count:
          type: integer
        tags:
          type: array
          items:
            type: string
...